Use boost::optional to store m_busyPeriodPacketUid
This is a minor nitpick. PacketGUID can be 0 (in fact, it is
usually 0 for the first simulated packet), so it is not correct
to use 0 to signal "no packet".
Instead, we hold the value in a boost::optional, so we have two
clear states: no packet (boost::none) or a packet with any
uint64_t guid value.