Correct elevation calculation
The angle was being calculated as the angle between the normal at
the ground station and the station to satellite vector. Then this
was being substracted from the normal (π/2) to obtain the elevation
from its orthogonal (tanget to ground body). Problem was that it
was using π/4 instead of π/2 as the normal.