Check that the solution crosses 0
The existing test just tested if the solution was very near to a
crossing point. However, sometimes it did find solutions where the
object was very near to the desired target, but without ever
crossing it. The new test should fix this, as it is expected that
there is a crossing point near (±1 second) from the solution.